nlp/requirements.txt
transformers>=4.37.0
torch>=2.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
neo4j>=5.14.0
